1차 수정 1. Calculator 메서드 분리 2. 테스트 분리
