+++ src/test/java/Calculator2Test.java
@@ -1,15 +1,15 @@
-import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class CalculatorTest {
+public class Calculator2Test {
 
-    private Calculator calculator = new Calculator();
+    private Calculator2 calculator = new Calculator2();
 
     @Test
-    @DisplayName("계산기1 정상 입력 테스트")
+    @DisplayName("계산기2 정상 입력 테스트")
     void calculate1_Normal_Test() {
         assertThat(calculator.calculate("1 + 2")).isEqualTo(3.0);
         assertThat(calculator.calculate("1")).isEqualTo(1.0);
@@ -17,7 +17,7 @@ public class CalculatorTest {
     }
 
     @Test
-    @DisplayName("계산기1 비정상 입력 테스트 (숫자 연속 입력)")
+    @DisplayName("계산기2 비정상 입력 테스트 (숫자 연속 입력)")
     void calculate1_Abnormal_Test1() {
         assertThrows(IllegalArgumentException.class, () -> {
             calculator.calculate("1 + 1 1 + 1");
@@ -25,7 +25,7 @@ public class CalculatorTest {
     }
 
     @Test
-    @DisplayName("계산기1 비정상 입력 테스트 (연산자 연속 입력)")
+    @DisplayName("계산기2 비정상 입력 테스트 (연산자 연속 입력)")
     void calculate1_Abnormal_Test2() {
         assertThrows(IllegalArgumentException.class, () -> {
             calculator.calculate("1 + - 1 + 1");
@@ -33,7 +33,7 @@ public class CalculatorTest {
     }
 
     @Test
-    @DisplayName("계산기1 비정상 입력 테스트 (연산자로 시작)")
+    @DisplayName("계산기2 비정상 입력 테스트 (연산자로 시작)")
     void calculate1_Abnormal_Test3() {
         assertThrows(IllegalArgumentException.class, () -> {
             calculator.calculate("+ 1");
@@ -41,7 +41,7 @@ public class CalculatorTest {
     }
 
     @Test
-    @DisplayName("계산기1 비정상 입력 테스트 (연산자로 종료)")
+    @DisplayName("계산기2 비정상 입력 테스트 (연산자로 종료)")
     void calculate1_Abnormal_Test4() {
         assertThrows(IllegalArgumentException.class, () -> {
             calculator.calculate("1 + 1 +");
